[ugcube] MTK ファイル分離
